rename: search-bar 파일 input 폴더로 이동
