+++ src/shared/ui/input/search-bar.stories.ts
