+++ src/shared/ui/input/search-bar.tsx
